Disable registry in 0.19.0-rc1
The default in 0.19.0-rc1 is to enable the registry and use a replicated
log. Unfortunately, this requires manual configuration of the new
options: --work_dir and --quorum. For now, we will build the 0.19.0-rc1
package with the registry set to in_memory by default, to ensure mesos
can still start without manual configuration. The mesos-master registry
default setting may change during the 0.19.0-rcX release process.