Fix cp failure when src dir not present
When cp is used with an expanded eggs[@] array, it will exit with an
error code if one of the directories does not exist, which will cause
the bash script to fail. Instead of relying on this approach to find the
eggs, just use a default python_dist path which is overridden if eggs
are found in the old location.