Correct separation of multiple attributes/resources
Hey guys,

when I'm using mesos-init-wrapper as it is on Ubuntu(14.04 - upstart), then I can't start the mesos-slave.
If we want to use a semicolon for delimiting multiple attributes/resources then we have to put whole argument between quotes, otherwise the second argument (right after the semicolon) is evaluated as a new command by Bash. Another approach to this problem is using a comma instead of semicolon on lines 97 and 105. Then we don't need quotation of whole attributes/resources arguments and it's closer to what the mesos-slave documentation says.

Cheers,
R.